requirements_init.txt
django-filter==2.1.0
django-filters==0.2.1
django-google-charts==0.1.1
django-qsstats-magic==1.0.0
django-tempus-dominus==5.1.2.4
djangorestframework==3.9.2
djangorestframework-filter==0.2.2
Markdown==3.0.1
paho-mqtt==1.4.0
pysqlite3==0.2.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
six==1.12.0
django-crispy-forms